% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/big_combinations.R
\name{big_combinations}
\alias{big_combinations}
\title{Combinations formula for large N}
\usage{
big_combinations(r, n)
}
\arguments{
\item{r}{Combination size}

\item{n}{Number of objects}
}
\description{
Implements an appoximation of the standard combinations formula using logfactorials
}

