% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/drop_trailing_substring.R
\name{drop_trailing_substring}
\alias{drop_trailing_substring}
\title{Drop trailing substring}
\usage{
drop_trailing_substring(input, substrings)
}
\arguments{
\item{input}{A character vector of text to be processed}

\item{substrings}{A character vector of words to be dropped from elements of input}
}
\description{
Drop trailing substring
}

