% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gg_footnote.R
\name{gg_footnote}
\alias{gg_footnote}
\title{Add a footnote to a ggplot2 plot}
\usage{
gg_footnote(footnote.text = format(Sys.time(), "\%d \%b \%Y"), size = 0.7,
  colour = grey(0.5))
}
\arguments{
\item{footnote.text}{Text for the footnote. Defaults to the current date.}

\item{size}{Size of footnote text}

\item{colour}{Colour of footnote text}
}
\description{
Adds a footnote to the ggplot2 plot in the current graphics device using the grid package.
}

