% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/multiplot.R
\name{multiplot}
\alias{multiplot}
\title{Multiple plot function
Developed by Winston Chang}
\usage{
multiplot(..., plotlist = NULL, file, cols = 1, layout = NULL)
}
\arguments{
\item{...}{ggplot objects to be plotted}

\item{plotlist}{an optional list of ggplot objects to be plotted}

\item{cols}{Number of columns in layout.}

\item{layout}{A matrix specifying the layout. If present, 'cols' is ignored. If the layout is something like matrix(c(1,2,3,3), nrow=2, byrow=TRUE), then plot 1 will go in the upper left, 2 will go in the upper right, and 3 will go all the way across the bottom.}
}
\description{
Multiple plot function
Developed by Winston Chang
}

