% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/print_bibentries.R
\name{print_bibentries}
\alias{print_bibentries}
\title{Manually print BibTex entries via knitcitations}
\usage{
print_bibentries(bibentries, biblist, ...)
}
\arguments{
\item{bibentries}{A character vector of Bibtex entry codes}

\item{biblist}{A BibEntry object (containing bibentries)}

\item{bibformat}{(optional) Output format, e.g. 'text' or 'pandoc'. Defaults to 'pandoc'}
}
\value{
A formatted bibliography to be processed by Pandoc
}
\description{
Manually print BibTex entries via knitcitations
}

