% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/replace_accented.R
\name{replace_accented}
\alias{replace_accented}
\title{Replace Accented Characters}
\usage{
replace_accented(text)
}
\arguments{
\item{text}{Character vector of text as input.}
}
\description{
Replaces accented characters from non-English language documents.
}

