% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/trim_whitespace.R
\name{trim_whitespace}
\alias{trim_whitespace}
\title{Trim Whitespace}
\usage{
trim_whitespace(input)
}
\arguments{
\item{input}{A character vector of input text.}
}
\description{
Trim Whitespace
}

